finished uploadMedia feature
